Dwarf Shop Etc
Defined the location of the Dwarf shop on Volcano Dungeon 5.
Defined two signs in town as decorations.
Removed the top and bottom tiles of the Joja Mart shop counter so that Accessible Tiles auto pathing won't navigate to the south / north end of the counter, as those tiles will only let you interact with them from the right.