Change Log.Error to Log.Trace for File Not Found in JsonLoader
Switched the logging level from `Log.Error` to `Log.Trace` when a file is not found in `JsonLoader`. This change allows the calling code to determine if the absence of a file is an error. Previously, an error message was generated when a user file failed to load, which wasn't always appropriate. Now using `Log.Trace` for better debugging flexibility.