Update Translator to split bracketed text from translationKey before lookup
Previously, "static_tile-common-stool[1]" was being looked up as is, not being found, and spoken as is. Now, the Translate functions will split bracketed text from the end of a key, lookup the key, and append the text from inside the brackets to the returned string. So "static_tile-common-stool[1]" will be properly spoken as "stool 1".