Fix `AccessibleLocation` not adding static tiles to categories
When an AccessibleLocation was initialized with an entire layer, I.E. for static tiles, the tiles in that layer were not being added to the CategoryTileMap. Thus we could not retrieve static tiles by category. Fixed by splitting out the logic to add tiles to the `CategoryTileMap` to a separate function which can also be used during initialization.