✨ Track cabin door and mail boxes and prepend owner's name
The owner name is pre-pended to the cabin/farm house only if it is not of the current owner. So, for the second player, the cabin's mail box will be read as `Mail Box` while the farm house's mail box will be read as `[First player's name] Mail Box` without the unread mail count status.