Add More Excluded Menus to IClickableMenuPatch
Add title screen menus to excluded menus in IClickableMenuPatch. Weird bug crashing the game in the CharacterCreation menu with a buffer overflow. The error message would not even be captured in the smapi log. It referenced the IClickableMenuPatch, which is unnecessary on CharacterCreation menu. Bug seems to no longer be reproduceable after excluding.