✨ Detect pet type in pet adoption shop menu
It's basically a shop menu but with pet licenses, and each license is
displayed with the specific pet and I was able to get the pet's breed
info which I used to get the pet's name from the existing translation
entrues.

I've also added entry for the turtles in
`character_creation_menu.en.ftl`, there are only two at the moment,
green turtle and purple turtle. #221